refactor(app): fix citation backwards-dependency on route folder
The citation subsystem (citation/reference-list.tsx, citation/how-to-cite.tsx)
imported the FormattedReference type and Reference component from a
page-specific route folder (app/scat-cs/inspect/*), a backwards dependency
from a reusable domain module into route-specific code.

app/scat-cs/inspect/cite.ts also byte-for-byte duplicated
reference2bibliography()/getReferenceLabel() already defined in
citation/cite.ts.

- Add citation/formatted-reference.ts: FormattedReference type +
  formatReference(), now importing reference2bibliography from citation/cite
  instead of duplicating it.
- Add citation/formatted-reference-link.tsx: the pre-formatted-reference
  link component, renamed to FormattedReferenceLink to avoid colliding with
  citation/reference.tsx's existing Reference component (which takes a raw
  ReferenceRecord, a different shape).
- Delete app/scat-cs/inspect/{cite.ts,types.ts,reference.tsx}; update all
  consumers to import from citation/ instead.
