--- app/src/app/scat-cs/inspect/types.ts
@@ -1,9 +1,0 @@
-// SPDX-FileCopyrightText: LXCat team
-//
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-export interface FormattedReference {
-  id: string;
-  ref: string;
-  url?: string;
-}
